Isolated the registry API from the auditor
