+++ pkg/logs/auditor/auditor_test.go
@@ -110,79 +110,6 @@ func (suite *AuditorTestSuite) TestAuditorCleansupRegistry() {
 	suite.Equal("43", suite.a.registry[otherpath].Offset)
 }
 
-func (suite *AuditorTestSuite) TestAuditorUnmarshalRegistryV0() {
-	input := `{
-	    "Registry": {
-	        "path1.log": {
-	            "Offset": 1,
-	            "Path": "path1.log",
-	            "Timestamp": "2006-01-12T01:01:01.000000001Z"
-	        },
-	        "path2.log": {
-	            "Offset": 2,
-	            "Path": "path2.log",
-	            "Timestamp": "2006-01-12T01:01:02.000000001Z"
-	        }
-	    },
-	    "Version": 0
-	}`
-	r, err := suite.a.unmarshalRegistry([]byte(input))
-	suite.Nil(err)
-	suite.Equal("1", r["file:path1.log"].Offset)
-	suite.Equal(1, r["file:path1.log"].LastUpdated.Second())
-	suite.Equal("2", r["file:path2.log"].Offset)
-	suite.Equal(2, r["file:path2.log"].LastUpdated.Second(), 2)
-}
-
-func (suite *AuditorTestSuite) TestAuditorUnmarshalRegistryV1() {
-	input := `{
-	    "Registry": {
-	        "path1.log": {
-	            "Offset": 1,
-	            "LastUpdated": "2006-01-12T01:01:01.000000001Z",
-	            "Timestamp": ""
-	        },
-	        "path2.log": {
-	            "Offset": 0,
-	            "LastUpdated": "2006-01-12T01:01:02.000000001Z",
-	            "Timestamp": "2006-01-12T01:01:03.000000001Z"
-	        }
-	    },
-	    "Version": 1
-	}`
-	r, err := suite.a.unmarshalRegistry([]byte(input))
-	suite.Nil(err)
-	suite.Equal("1", r["path1.log"].Offset)
-	suite.Equal(1, r["path1.log"].LastUpdated.Second())
-
-	suite.Equal("2006-01-12T01:01:03.000000001Z", r["path2.log"].Offset)
-	suite.Equal(2, r["path2.log"].LastUpdated.Second(), 2)
-}
-
-func (suite *AuditorTestSuite) TestAuditorUnmarshalRegistryV2() {
-	input := `{
-	    "Registry": {
-	        "path1.log": {
-	            "Offset": "1",
-	            "LastUpdated": "2006-01-12T01:01:01.000000001Z"	            
-	        },
-	        "path2.log": {
-	            "Offset": "2006-01-12T01:01:03.000000001Z",
-	            "LastUpdated": "2006-01-12T01:01:02.000000001Z"	            
-	        }
-	    },
-	    "Version": 2
-	}`
-	r, err := suite.a.unmarshalRegistry([]byte(input))
-	suite.Nil(err)
-
-	suite.Equal("1", r["path1.log"].Offset)
-	suite.Equal(1, r["path1.log"].LastUpdated.Second())
-
-	suite.Equal("2006-01-12T01:01:03.000000001Z", r["path2.log"].Offset)
-	suite.Equal(2, r["path2.log"].LastUpdated.Second())
-}
-
 func TestScannerTestSuite(t *testing.T) {
 	suite.Run(t, new(AuditorTestSuite))
 }
